Add: Provider, update: tabs page
Installing provider in the pubspec.
Update tabs page with a singleton to listen the actual page selected by the user.
Wrapping the scaffold with a Change Notifier Provider by this way create a global instance of provider.
Instancing and initializing a new property into the bottom navigation bar to avoid the provider to find the instance of NavegacionModel into the widget tree.
Using the new property in the current index and in the onTap to listen if the user change the bottom navigation bar item and if there is a change, refresh the page and show the change into the screen.