Update: tabs page.
Updating tabs page with new property into the PageView widget by this way avoid the app the manage the controller.
Adding a new property into the Navegacion Model widget to manage the page controller.
Adding the getter of the page controller.
Adding the animate to page into the setter of the paginaActual by this way manage the page view depends on where the user selection.