Update: Tab1 page, update: tabs page, update: news service.
Updating tab1 page with the instance of provider and a container.
Updating tabs page to show the tab1 page instead a container.
Updating news service with the url and its uri, authority, unencodedPath and the queryParameters, also the news response from json with the resp.body to catch all the articles from the API.