Update: Lista noticias, update: main
Updating lista noticias with 3 new classes, one for the top bar that counts the number and the source of the news, the second one is for the title, and the third one is to implement the image of the new.
Updating main with and http overrides to avoid all certificates.