Add: Assets, update: lista noticias.
Adding assets images to use in the placeholder and to show a no image if the new has no image.
Updating lista noticias with the implementation of the image from the url, making a condition if the url is empty show an asset image, also adding the widget to show the description of the new.