Update: tab1 page, update: tab2 page, update: tabs page
Updating tab1 page converting from stateless widget to a statefull widget to use the Automatic Keep Alive Client Mixin to keep the state if the user change the page.
Updating tab2 page with a stateless widget and get ready the page to implement the encabezados.
Updating tabs page, implementing the tab2 page into the page view.