Add: Category model, update: tab2 page, update: news service, add: font awesome.
Adding category model to create the constructor to get the icon and the name of the news category
Updating tab2 page to implement the model created in category model, using a provider and a list view builder with a horizontal scroll direction.
Update news service with a list of categories with its icon, that will be use in the tab2 page.