Update: tab2 page, update: news service
Updating tab2 page with an extract of the icon widget to simplify the code into the list view builder, adding shape and color to the icon to enhance the visual looking of the app also implement the gesture detector to soon add the on tap function and redirect the page depends on what the user category wants.
Updating tab2 page changing the icon.