Update: news service, update: tab2 page
Updating the news service creating an empty map to receive the name and the articles list.
Ading a foreach to wander all the categories list.
Adding a getter and setter of the selected category.
Finally adding a method to manage the info from the API.
Update tab2 page, specifically into the gesture detector into its onTap method, creating a propertie of provider to read when the user taps in an icon of the tab2 page.