Update: news service, update: tab2 page.
Updating news service with a getter to know what is the category selected by the user, also creating a boolean propertie to know if the page is loading or is already loaded.
Updating tab2 page with a propertie of provider inside the tab2 page and creating a new expanded widget to show all the articles.